backend/requirements.txt
fastapi[standard]
sqlmodel
sqlalchemy
pydantic
pyodbc
PyJWT
bcrypt